# Use multiple threads to import HDR images