… thirdparty/cvtt/ConvectionKernels_BC67.cpp
… thirdparty/cvtt/patches/revert_BC6H_reorg.patch