### Part1: Performance

Fixes: https://github.com/godotengine/godot/issues/61948
Regression from: https://github.com/godotengine/godot/pull/49775

https://github.com/godotengine/godot/pull/49775 disabled threads by forcing ``num_threads`` to always be 0. Then https://github.com/godotengine/godot/pull/63365 removed all the multithreading code entirely because ``num_threads`` was always 0.

This PR restores multithreading to the BCPT import path (used by HDR images) which brings performance back to what it was in Alpha 1.

| Build             | 1k texture  | 2k texture | 15k texture |
|-------------------|-------------|------------|------------|
| Master (debug)    | 270 seconds |            |             |
| This PR (debug)   | 60 seconds  |            |             |
| Master (release)  | 31 seconds  |            |             |
| This PR (release) | 4.5 seconds |            |             |
| RC3               |             | 75 seconds | *            |
| This PR (release) |             | 15 seconds |13 minutes |

_* I won't test the 15k texture on RC3 and you can't make me_

### Part 2: Quality

In addition to a performance regression, https://github.com/godotengine/godot/pull/49775 also introduced a quality regression.

The quality issue appears to come from an [upstream change](https://github.com/elasota/cvtt/commit/2e4b6b2747aec11f4cc6dd09ef43fa8ce769f6e2) that was included in #49775. So I have removed the changes made in upstream that directly affect BC6H. The changes to BC7 I left alone as they appear to be fine. 

_Before https://github.com/godotengine/godot/pull/49775_
![Screenshot from 2023-02-22 12-13-25](https://user-images.githubusercontent.com/16521339/220752297-27d8d925-7c00-40df-ad0f-256f38db8317.png)

_After https://github.com/godotengine/godot/pull/49775_
![Screenshot from 2023-02-22 12-14-32](https://user-images.githubusercontent.com/16521339/220752293-f751d272-010d-4aac-b982-2971ace98f65.png)

_Alpha 1_
![Screenshot from 2023-02-22 13-04-39](https://user-images.githubusercontent.com/16521339/220758855-ae511d78-5db2-4747-8bb2-21afcc45ee39.png)


A helpful MRP for testers. This contains the 1024x512 HDR file I used for testing
[HDR-import.zip](https://github.com/godotengine/godot/files/10799661/HDR-import.zip)