@Calinou 2023-02-21
The import time in debug is still pretty slow. Could we forcibly build CVTT with release optimizations like with CVTT?