@elasota 2023-10-29
FYI the BC6H decoder regression should be fixed upstream: https://github.com/elasota/ConvectionKernels/commit/350416daa4e98f1c17ffc273b134d0120a2ef230