@clayjohn 2023-10-29
> FYI the BC6H decoder regression should be fixed upstream: [elasota/ConvectionKernels@350416d](https://github.com/elasota/ConvectionKernels/commit/350416daa4e98f1c17ffc273b134d0120a2ef230)

Amazing! Thank you. Its a bit awe inspiring to see how few lines of code it took to fix the issue. I am admittedly totally lost with this stuff.